refactor: Expose public API and group functionality into extension for easier reading and navigation
