+++ Sources/FeaturevisorSDK/Instance+Status.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public extension FeaturevisorInstance {
+
+    // MARK: - Statuses
+
+    func isReady() -> Bool {
+        return statuses.ready
+    }
+
+}
